+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.0.3-alpha.11",
-  "contentSha256_12": "5ed9482df8a8",
+  "pkgVersion": "1.0.3-alpha.12",
+  "contentSha256_12": "f7f5c90c5495",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.0.3-alpha.11";
-export const contentSha256_12 = "5ed9482df8a8";
+export const pkgVersion = "1.0.3-alpha.12";
+export const contentSha256_12 = "f7f5c90c5495";
 export const file = "nodalia-cards.bundle.js";